% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_adtte_format.R
\name{check_adtte_format}
\alias{check_adtte_format}
\title{Check required fields in ADTTE}
\usage{
check_adtte_format(adtte)
}
\arguments{
\item{adtte}{A data frame of ADTTE}
}
\value{
TRUE if all required fields are present, otherwise stop with message
}
\description{
Check required fields in ADTTE
}
