% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_surv_data.R
\name{prepare_surv_data}
\alias{prepare_surv_data}
\title{Prepare survival analysis data from ADSL and ADTTE}
\usage{
prepare_surv_data(adsl, adtte, paramcd = "OS")
}
\arguments{
\item{adsl}{A data frame containing ADSL data}

\item{adtte}{A data frame containing ADTTE data}

\item{paramcd}{Parameter code to filter in ADTTE (default: "OS")}
}
\value{
A merged data frame ready for survival::Surv()
}
\description{
Prepare survival analysis data from ADSL and ADTTE
}
