% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_adsl}
\alias{simulate_adsl}
\title{Simulate ADSL dataset (mock CDISC)}
\usage{
simulate_adsl(n = 10)
}
\arguments{
\item{n}{Number of subjects}
}
\value{
A data frame with simulated ADSL variables
}
\description{
Simulate ADSL dataset (mock CDISC)
}
