% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_adtte}
\alias{simulate_adtte}
\title{Simulate ADTTE dataset (mock CDISC)}
\usage{
simulate_adtte(n = 10, paramcd = "OS")
}
\arguments{
\item{n}{Number of rows}

\item{paramcd}{PARAMCD value (e.g., "OS")}
}
\value{
A data frame with simulated ADTTE variables
}
\description{
Simulate ADTTE dataset (mock CDISC)
}
